Оптимизировал роуты, добавив контроллеры для операций с постами
